Revert "Merge pull request #176 from SquadDevelop/feat/134-ai"
This reverts commit c49587c9cf8c6f6acb9cdd88814e5460adfb63da, reversing
changes made to 7cc7d1adf85637dae37eb681feef17c6dc06da8a.

+++ src/main/java/com/project/baedalsodae/review/dto/request/ReviewRequest.java
@@ -1,3 +1,3 @@
 package com.project.baedalsodae.review.dto.request;
 
-public record ReviewRequest(double rating, String comment) {}
+public record ReviewRequest(int rating, String comment) {}
